Вынес сетевые взаимодействия в отдельный файл
