+++ cli.py
@@ -34,10 +34,10 @@ else:
     print('"https://github.com/Vi-812/git" либо "Vi-812/git"')
     sys.exit()
 
-xz = analyzer.GitGraphql(repository_owner, repository_name)
-# xz = analyzer.GitGraphql('facebook', 'jest')
-# xz = analyzer.GitGraphql('vi-812', 'empty')
-# xz = analyzer.GitGraphql('pallets', 'flask')
+# xz = analyzer.GitHubAnalyz(repository_owner, repository_name)
+xz = analyzer.GitHubAnalyz('facebook', 'jest')
+# xz = analyzer.GitHubAnalyz('vi-812', 'empty')
+# xz = analyzer.GitHubAnalyz('pallets', 'flask')
 xz.get_info_labels()
 xz.get_bug_issues()
 xz.analyz_bug_issues()
